# Conflicts:
#	src/core/lib/core_data_model/CMakeLists.txt
#	src/core/lib/core_data_model/i_item_role.hpp
#	src/core/lib/core_generic_plugin_manager/generic_plugin_manager.cpp
#	src/core/lib/core_qt_common/CMakeLists.txt
#	src/core/lib/core_qt_common/shared_controls.cpp

+++ src/core/testing/plg_ui_main_test/pages/test_page.hpp
@@ -30,8 +30,8 @@ private:
 	void setTextField( const std::wstring & text );
 	void getTextField( std::wstring * text ) const;
 
-	void setSlideData( const int & length );
-	void getSlideData( int * length ) const;
+	void setSlideData(const double & length);
+	void getSlideData(double * length) const;
 
 	static int getSlideMaxData();
 	static int getSlideMinData();
@@ -90,7 +90,7 @@ private:
 	bool bChecked_;
 	bool boolTest_;
 	std::wstring text_;
-	int curSlideData_;
+	double curSlideData_;
 	int curNum_;
 	int curSelected_;
 	int enumValue_;
